requirements.txt
flask-restful==0.3.6
flask-sqlalchemy==2.3.2
beautifulsoup4==4.6.0
requests==2.18.4
pyodbc==4.0.23
